Extract common template classes and deprecate polymer support
