+++ flow-server/src/main/java/com/vaadin/flow/component/polymertemplate/NpmTemplateParser.java
@@ -30,6 +30,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.vaadin.flow.component.dependency.JsModule;
+import com.vaadin.flow.component.littemplate.LitTemplate;
 import com.vaadin.flow.function.DeploymentConfiguration;
 import com.vaadin.flow.internal.AnnotationReader;
 import com.vaadin.flow.internal.Pair;
@@ -56,7 +57,11 @@ import elemental.json.JsonObject;
  * @since 2.0
  *
  * @see BundleParser
+ * 
+ * @deprecated Use {@code LitTemplateParserImpl} class for {@link LitTemplate}
+ *             components, polymer templates are deprecated
  */
+@Deprecated
 public class NpmTemplateParser implements TemplateParser {
 
     private static final TemplateParser INSTANCE = new NpmTemplateParser();
@@ -66,8 +71,9 @@ public class NpmTemplateParser implements TemplateParser {
     private JsonObject jsonStats;
 
     /**
-     * The default constructor. Protected in order to prevent direct instantiation,
-     * but not private in order to allow mocking/overrides for testing purposes.
+     * The default constructor. Protected in order to prevent direct
+     * instantiation, but not private in order to allow mocking/overrides for
+     * testing purposes.
      */
     protected NpmTemplateParser() {
     }
@@ -84,8 +90,8 @@ public class NpmTemplateParser implements TemplateParser {
         List<Dependency> dependencies = AnnotationReader
                 .getAnnotationsFor(clazz, JsModule.class).stream()
                 .map(jsModule -> new Dependency(Dependency.Type.JS_MODULE,
-                        jsModule.value(),
-                        LoadMode.EAGER)) // load mode doesn't matter here
+                        jsModule.value(), LoadMode.EAGER)) // load mode doesn't
+                                                           // matter here
                 .collect(Collectors.toList());
 
         for (DependencyFilter filter : service.getDependencyFilters()) {
@@ -159,10 +165,14 @@ public class NpmTemplateParser implements TemplateParser {
     /**
      * Finds the JavaScript sources for given tag.
      *
-     * @param tag the value of the {@link com.vaadin.flow.component.Tag} annotation,
-     *            e.g. `my-component`
-     * @param url the URL resolved according to the {@link com.vaadin.flow.component.dependency.JsModule}
-     *            spec, for example {@code ./view/my-view.js} or {@code @vaadin/vaadin-button.js}.
+     * @param tag
+     *            the value of the {@link com.vaadin.flow.component.Tag}
+     *            annotation, e.g. `my-component`
+     * @param url
+     *            the URL resolved according to the
+     *            {@link com.vaadin.flow.component.dependency.JsModule} spec,
+     *            for example {@code ./view/my-view.js} or
+     *            {@code @vaadin/vaadin-button.js}.
      * @return the .js source which declares given custom element, or null if no
      *         such source can be found.
      */
